requirements.txt
google==2.0.1
protobuf==3.6.0
pycrypto==2.6.1
redis==2.10.6